+++ storage/log_iterator.go
@@ -8,7 +8,6 @@ import (
 	"runtime"
 	"sync"
 
-	"github.com/evergreen-ci/logkeeper/model"
 	"github.com/evergreen-ci/pail"
 	"github.com/mongodb/grip"
 	"github.com/mongodb/grip/message"
@@ -21,14 +20,14 @@ import (
 type LogIterator interface {
 	Iterator
 	// Item returns the current LogLine item held by the iterator.
-	Item() model.LogLineItem
+	Item() LogLineItem
 	// Reverse returns a reversed copy of the iterator.
 	Reverse() LogIterator
 	// IsReversed returns true if the iterator is in reverse order and
 	// false otherwise.
 	IsReversed() bool
 	// Stream returns a chan of log lines from the iterator.
-	Stream(context.Context) chan *model.LogLineItem
+	Stream(context.Context) chan *LogLineItem
 }
 
 //////////////////////
@@ -45,7 +44,7 @@ type serializedIterator struct {
 	currentReadCloser    io.ReadCloser
 	currentReverseReader *reverseLineReader
 	currentReader        *bufio.Reader
-	currentItem          model.LogLineItem
+	currentItem          LogLineItem
 	catcher              grip.Catcher
 	exhausted            bool
 	closed               bool
@@ -136,6 +135,8 @@ func (i *serializedIterator) Next(ctx context.Context) bool {
 			i.catcher.Wrap(err, "parsing timestamp")
 			return false
 		}
+		item.Global = i.chunks[i.keyIndex].TestID != ""
+
 		i.lineCount++
 
 		if item.Timestamp.After(i.timeRange.EndAt) && !i.reverse {
@@ -161,7 +162,7 @@ func (i *serializedIterator) Exhausted() bool { return i.exhausted }
 
 func (i *serializedIterator) Err() error { return i.catcher.Resolve() }
 
-func (i *serializedIterator) Item() model.LogLineItem { return i.currentItem }
+func (i *serializedIterator) Item() LogLineItem { return i.currentItem }
 
 func (i *serializedIterator) Close() error {
 	i.closed = true
@@ -172,7 +173,7 @@ func (i *serializedIterator) Close() error {
 	return nil
 }
 
-func (i *serializedIterator) Stream(ctx context.Context) chan *model.LogLineItem {
+func (i *serializedIterator) Stream(ctx context.Context) chan *LogLineItem {
 	return streamFromLogIterator(ctx, i)
 }
 
@@ -192,7 +193,7 @@ type batchedIterator struct {
 	readers              map[string]io.ReadCloser
 	currentReverseReader *reverseLineReader
 	currentReader        *bufio.Reader
-	currentItem          model.LogLineItem
+	currentItem          LogLineItem
 	catcher              grip.Catcher
 	exhausted            bool
 	closed               bool
@@ -380,7 +381,7 @@ func (i *batchedIterator) Exhausted() bool { return i.exhausted }
 
 func (i *batchedIterator) Err() error { return i.catcher.Resolve() }
 
-func (i *batchedIterator) Item() model.LogLineItem { return i.currentItem }
+func (i *batchedIterator) Item() LogLineItem { return i.currentItem }
 
 func (i *batchedIterator) Close() error {
 	i.closed = true
@@ -393,7 +394,7 @@ func (i *batchedIterator) Close() error {
 	return catcher.Resolve()
 }
 
-func (i *batchedIterator) Stream(ctx context.Context) chan *model.LogLineItem {
+func (i *batchedIterator) Stream(ctx context.Context) chan *LogLineItem {
 	return streamFromLogIterator(ctx, i)
 }
 
@@ -404,7 +405,7 @@ func (i *batchedIterator) Stream(ctx context.Context) chan *model.LogLineItem {
 type mergingIterator struct {
 	iterators    []LogIterator
 	iteratorHeap *LogIteratorHeap
-	currentItem  model.LogLineItem
+	currentItem  LogLineItem
 	catcher      grip.Catcher
 	started      bool
 }
@@ -491,7 +492,7 @@ func (i *mergingIterator) init(ctx context.Context) {
 
 func (i *mergingIterator) Err() error { return i.catcher.Resolve() }
 
-func (i *mergingIterator) Item() model.LogLineItem { return i.currentItem }
+func (i *mergingIterator) Item() LogLineItem { return i.currentItem }
 
 func (i *mergingIterator) Close() error {
 	catcher := grip.NewBasicCatcher()
@@ -507,7 +508,7 @@ func (i *mergingIterator) Close() error {
 	return catcher.Resolve()
 }
 
-func (i *mergingIterator) Stream(ctx context.Context) chan *model.LogLineItem {
+func (i *mergingIterator) Stream(ctx context.Context) chan *LogLineItem {
 	return streamFromLogIterator(ctx, i)
 }
 
@@ -532,8 +533,8 @@ func reverseChunks(chunks []LogChunkInfo) {
 	}
 }
 
-func streamFromLogIterator(ctx context.Context, iter LogIterator) chan *model.LogLineItem {
-	logLines := make(chan *model.LogLineItem)
+func streamFromLogIterator(ctx context.Context, iter LogIterator) chan *LogLineItem {
+	logLines := make(chan *LogLineItem)
 	go func() {
 		defer recovery.LogStackTraceAndContinue("streaming lines from log iterator")
 		defer close(logLines)
