+++ views.go
@@ -9,14 +9,11 @@ import (
 	"sync"
 	"time"
 
-	"github.com/evergreen-ci/logkeeper/env"
-	"github.com/evergreen-ci/logkeeper/model"
 	"github.com/evergreen-ci/logkeeper/storage"
 	"github.com/evergreen-ci/render"
 	"github.com/evergreen-ci/utility"
 	"github.com/gorilla/handlers"
 	"github.com/gorilla/mux"
-	"github.com/mongodb/amboy"
 	"github.com/mongodb/grip"
 	"github.com/mongodb/grip/message"
 	"github.com/mongodb/grip/recovery"
@@ -97,9 +94,9 @@ type apiError struct {
 }
 
 type logFetchResponse struct {
-	logLines chan *model.LogLineItem
-	build    *model.Build
-	test     *model.Test
+	logLines chan *storage.LogLineItem
+	build    *storage.Build
+	test     *storage.Test
 }
 
 func (lk *logKeeper) logErrorf(r *http.Request, format string, v ...interface{}) {
@@ -142,7 +139,7 @@ func (lk *logKeeper) createBuild(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	id, err := model.NewBuildID(payload.Builder, payload.BuildNum)
+	id, err := storage.NewBuildID(payload.Builder, payload.BuildNum)
 	if err != nil {
 		lk.logErrorf(r, "creating new build ID: %v", err)
 		lk.render.WriteJSON(w, http.StatusInternalServerError, apiError{Err: "creating new build ID"})
@@ -175,7 +172,7 @@ func (lk *logKeeper) createTest(w http.ResponseWriter, r *http.Request) {
 	defer r.Body.Close()
 
 	buildID := mux.Vars(r)["build_id"]
-	testID := string(model.NewTestID(time.Now()))
+	testID := string(storage.NewTestID(time.Now()))
 
 	if err := lk.checkContentLength(r); err != nil {
 		lk.logErrorf(r, "content length limit exceeded for create test: %s", err.Err)
@@ -252,7 +249,7 @@ func (lk *logKeeper) appendGlobalLog(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var lines []model.LogLine
+	var lines []storage.LogLineItem
 	if err := readJSON(r.Body, lk.opts.MaxRequestSize, &lines); err != nil {
 		lk.logErrorf(r, "bad request to append log lines to build '%s': %s", buildID, err.Err)
 		lk.render.WriteJSON(w, err.code, err)
@@ -262,7 +259,7 @@ func (lk *logKeeper) appendGlobalLog(w http.ResponseWriter, r *http.Request) {
 		lk.render.WriteJSON(w, http.StatusOK, "")
 		return
 	}
-	chunks, err := model.GroupLines(lines, maxLogBytes)
+	chunks, err := storage.GroupLines(lines, maxLogBytes)
 	if err != nil {
 		lk.logErrorf(r, "unmarshalling log lines for build '%s': %v", buildID, err)
 		lk.render.WriteJSON(w, http.StatusBadRequest, apiError{Err: err.Error()})
@@ -308,7 +305,7 @@ func (lk *logKeeper) appendLog(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var lines []model.LogLine
+	var lines []storage.LogLineItem
 	if err := readJSON(r.Body, lk.opts.MaxRequestSize, &lines); err != nil {
 		lk.logErrorf(r, "bad request to append log to test '%s' for build '%s': %s", testID, buildID, err.Err)
 		lk.render.WriteJSON(w, err.code, err)
@@ -318,7 +315,7 @@ func (lk *logKeeper) appendLog(w http.ResponseWriter, r *http.Request) {
 		lk.render.WriteJSON(w, http.StatusOK, "")
 		return
 	}
-	chunks, err := model.GroupLines(lines, maxLogBytes)
+	chunks, err := storage.GroupLines(lines, maxLogBytes)
 	if err != nil {
 		lk.logErrorf(r, "unmarshalling log lines: %v", err)
 		lk.render.WriteJSON(w, http.StatusBadRequest, apiError{Err: err.Error()})
@@ -348,33 +345,28 @@ func (lk *logKeeper) viewBuild(w http.ResponseWriter, r *http.Request) {
 	buildID := vars["build_id"]
 
 	var (
-		build    *model.Build
-		tests    []model.Test
+		build    *storage.Build
+		tests    []storage.Test
 		fetchErr *apiError
 	)
 	build, tests, fetchErr = lk.viewBucketBuild(r, buildID)
-	if fetchErr != nil && fetchErr.code == http.StatusNotFound {
-		// If not found in the pail-backed offline storage, fallback to
-		// the DB.
-		build, tests, fetchErr = lk.viewDBBuild(r, buildID)
-	}
 	if fetchErr != nil {
 		lk.render.WriteJSON(w, fetchErr.code, *fetchErr)
 		return
 	}
 
 	lk.render.WriteHTML(w, http.StatusOK, struct {
-		Build *model.Build
-		Tests []model.Test
+		Build *storage.Build
+		Tests []storage.Test
 	}{build, tests}, "base", "build.html")
 }
 
-func (lk *logKeeper) viewBucketBuild(r *http.Request, buildID string) (*model.Build, []model.Test, *apiError) {
+func (lk *logKeeper) viewBucketBuild(r *http.Request, buildID string) (*storage.Build, []storage.Test, *apiError) {
 	var (
 		wg       sync.WaitGroup
-		build    *model.Build
+		build    *storage.Build
 		buildErr error
-		tests    []model.Test
+		tests    []storage.Test
 		testsErr error
 	)
 
@@ -409,25 +401,6 @@ func (lk *logKeeper) viewBucketBuild(r *http.Request, buildID string) (*model.Bu
 	return build, tests, nil
 }
 
-func (lk *logKeeper) viewDBBuild(r *http.Request, buildID string) (*model.Build, []model.Test, *apiError) {
-	build, err := model.FindBuildById(buildID)
-	if err != nil {
-		lk.logErrorf(r, "finding DB build '%s': %v", buildID, err)
-		return nil, nil, &apiError{Err: "finding build", code: http.StatusInternalServerError}
-	}
-	if build == nil {
-		return nil, nil, &apiError{Err: "build not found", code: http.StatusNotFound}
-	}
-
-	tests, err := model.FindTestsForBuild(buildID)
-	if err != nil {
-		lk.logErrorf(r, "finding tests for DB build '%s': %v", buildID, err)
-		return nil, nil, &apiError{Err: "finding tests for build", code: http.StatusInternalServerError}
-	}
-
-	return build, tests, nil
-}
-
 ///////////////////////////////////////////////////////////////////////////////
 //
 // GET /build/{build_id}/all
@@ -444,17 +417,7 @@ func (lk *logKeeper) viewAllLogs(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var (
-		resp     *logFetchResponse
-		fetchErr *apiError
-		fallback bool
-	)
-	resp, fetchErr, fallback = lk.viewBucketLogs(r, buildID, "")
-	if fallback {
-		// If not found in the pail-backed offline storage, fallback to
-		// the DB.
-		resp, fetchErr = lk.viewAllDBLogs(r, buildID)
-	}
+	resp, fetchErr := lk.viewBucketLogs(r, buildID, "")
 	if fetchErr != nil {
 		lk.render.WriteJSON(w, fetchErr.code, *fetchErr)
 		return
@@ -472,13 +435,13 @@ func (lk *logKeeper) viewAllLogs(w http.ResponseWriter, r *http.Request) {
 		return
 	} else {
 		err := lk.render.StreamHTML(w, http.StatusOK, struct {
-			LogLines chan *model.LogLineItem
+			LogLines chan *storage.LogLineItem
 			BuildId  string
 			Builder  string
 			TestId   string
 			TestName string
-			Info     model.BuildInfo
-		}{resp.logLines, resp.build.Id, resp.build.Builder, "", "All logs", resp.build.Info}, "base", "test.html")
+			TaskID   string
+		}{resp.logLines, resp.build.ID, resp.build.Builder, "", "All logs", resp.build.TaskID}, "base", "test.html")
 		if err != nil {
 			lk.logErrorf(r, "rendering template: %v", err)
 		}
@@ -503,17 +466,7 @@ func (lk *logKeeper) viewTestLogs(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var (
-		resp     *logFetchResponse
-		fetchErr *apiError
-		fallback bool
-	)
-	resp, fetchErr, fallback = lk.viewBucketLogs(r, buildID, testID)
-	if fallback {
-		// If not found in the pail-backed offline storage, fallback to
-		// the DB.
-		resp, fetchErr = lk.viewDBTestLogs(r, buildID, testID)
-	}
+	resp, fetchErr := lk.viewBucketLogs(r, buildID, testID)
 	if fetchErr != nil {
 		lk.render.WriteJSON(w, fetchErr.code, *fetchErr)
 		return
@@ -535,27 +488,27 @@ func (lk *logKeeper) viewTestLogs(w http.ResponseWriter, r *http.Request) {
 		}
 	} else {
 		err := lk.render.StreamHTML(w, http.StatusOK, struct {
-			LogLines chan *model.LogLineItem
+			LogLines chan *storage.LogLineItem
 			BuildId  string
 			Builder  string
 			TestId   string
 			TestName string
-			Info     model.TestInfo
-		}{resp.logLines, resp.build.Id, resp.build.Builder, string(resp.test.Id), resp.test.Name, resp.test.Info}, "base", "test.html")
+			TaskID   string
+		}{resp.logLines, resp.build.ID, resp.build.Builder, string(resp.test.ID), resp.test.Name, resp.test.TaskID}, "base", "test.html")
 		if err != nil {
 			lk.logErrorf(r, "rendering template: %v", err)
 		}
 	}
 }
 
-func (lk *logKeeper) viewBucketLogs(r *http.Request, buildID string, testID string) (*logFetchResponse, *apiError, bool) {
+func (lk *logKeeper) viewBucketLogs(r *http.Request, buildID string, testID string) (*logFetchResponse, *apiError) {
 	var (
 		wg          sync.WaitGroup
-		build       *model.Build
+		build       *storage.Build
 		buildErr    error
-		test        *model.Test
+		test        *storage.Test
 		testErr     error
-		logLines    chan *model.LogLineItem
+		logLines    chan *storage.LogLineItem
 		logLinesErr error
 	)
 
@@ -585,75 +538,21 @@ func (lk *logKeeper) viewBucketLogs(r *http.Request, buildID string, testID stri
 
 	if buildErr != nil {
 		lk.logErrorf(r, "finding build '%s': %v", buildID, buildErr)
-		return nil, &apiError{Err: "finding build", code: http.StatusInternalServerError}, false
+		return nil, &apiError{Err: "finding build", code: http.StatusInternalServerError}
 	}
 	if build == nil {
-		return nil, &apiError{Err: "build not found", code: http.StatusNotFound}, true
+		return nil, &apiError{Err: "build not found", code: http.StatusNotFound}
 	}
 	if testErr != nil {
 		lk.logErrorf(r, "finding test '%s' for build '%s': %v", testID, buildID, testErr)
-		return nil, &apiError{Err: "finding test", code: http.StatusInternalServerError}, false
+		return nil, &apiError{Err: "finding test", code: http.StatusInternalServerError}
 	}
 	if testID != "" && test == nil {
-		return nil, &apiError{Err: "test not found", code: http.StatusNotFound}, false
+		return nil, &apiError{Err: "test not found", code: http.StatusNotFound}
 	}
 	if logLinesErr != nil {
 		lk.logErrorf(r, "downloading logs for build '%s': %v", buildID, logLinesErr)
-		return nil, &apiError{Err: "downloading logs", code: http.StatusInternalServerError}, false
-	}
-
-	return &logFetchResponse{
-		logLines: logLines,
-		build:    build,
-		test:     test,
-	}, nil, false
-}
-
-func (lk *logKeeper) viewAllDBLogs(r *http.Request, buildID string) (*logFetchResponse, *apiError) {
-	build, err := model.FindBuildById(buildID)
-	if err != nil {
-		lk.logErrorf(r, "finding build '%s': %v", buildID, err)
-		return nil, &apiError{Err: "finding build", code: http.StatusInternalServerError}
-	}
-	if build == nil {
-		return nil, &apiError{Err: "build not found", code: http.StatusNotFound}
-	}
-
-	logLines, err := model.AllLogs(build.Id)
-	if err != nil {
-		lk.logErrorf(r, "downloading all DB logs for build '%s': %v", buildID, err)
-		return nil, &apiError{Err: "downloading all logs", code: http.StatusInternalServerError}
-	}
-
-	return &logFetchResponse{
-		logLines: logLines,
-		build:    build,
-	}, nil
-}
-
-func (lk *logKeeper) viewDBTestLogs(r *http.Request, buildID string, testID string) (*logFetchResponse, *apiError) {
-	build, err := model.FindBuildById(buildID)
-	if err != nil {
-		lk.logErrorf(r, "finding build '%s': %v", buildID, err)
-		return nil, &apiError{Err: "finding build", code: http.StatusInternalServerError}
-	}
-	if build == nil {
-		return nil, &apiError{Err: "build not found", code: http.StatusNotFound}
-	}
-
-	test, err := model.FindTestByID(testID)
-	if err != nil {
-		lk.logErrorf(r, "finding test '%s' for build '%s': %v", testID, buildID, err)
-		return nil, &apiError{Err: "finding test", code: http.StatusInternalServerError}
-	}
-	if test == nil {
-		return nil, &apiError{Err: "test not found", code: http.StatusNotFound}
-	}
-
-	logLines, err := model.MergedTestLogs(test)
-	if err != nil {
-		lk.logErrorf(r, "finding logs for test '%s' in build '%s': %v", testID, buildID, err)
-		return nil, &apiError{Err: err.Error(), code: http.StatusInternalServerError}
+		return nil, &apiError{Err: "downloading logs", code: http.StatusInternalServerError}
 	}
 
 	return &logFetchResponse{
@@ -671,24 +570,13 @@ func (lk *logKeeper) checkAppHealth(w http.ResponseWriter, r *http.Request) {
 	defer r.Body.Close()
 
 	resp := struct {
-		Err             string           `json:"err"`
-		MaxRequestSize  int              `json:"maxRequestSize"`
-		DB              bool             `json:"db"`
-		Build           string           `json:"build_id"`
-		BatchSize       int              `json:"batch_size"`
-		NumWorkers      int              `json:"workers"`
-		DurationSeconds float64          `json:"dur_secs"`
-		CleanupStatus   amboy.QueueStats `json:"cleanup_queue_stats"`
+		Build          string `json:"build_id"`
+		MaxRequestSize int    `json:"maxRequestSize"`
 	}{
-		Build:           BuildRevision,
-		MaxRequestSize:  lk.opts.MaxRequestSize,
-		BatchSize:       CleanupBatchSize,
-		NumWorkers:      AmboyWorkers,
-		DurationSeconds: AmboyInterval.Seconds(),
-		CleanupStatus:   env.CleanupQueue().Stats(r.Context()),
+		Build:          BuildRevision,
+		MaxRequestSize: lk.opts.MaxRequestSize,
 	}
 
-	resp.DB = true
 	lk.render.WriteJSON(w, http.StatusOK, &resp)
 }
 
@@ -731,8 +619,6 @@ func (lk *logKeeper) NewRouter() *mux.Router {
 	r.StrictSlash(true).Path("/build/{build_id}/all").Methods("GET").Handler(handlers.CompressHandler(http.HandlerFunc(lk.viewAllLogs)))
 	r.StrictSlash(true).Path("/build/{build_id}/test/{test_id}").Methods("GET").Handler(handlers.CompressHandler(http.HandlerFunc(lk.viewTestLogs)))
 	r.PathPrefix("/lobster").Methods("GET").HandlerFunc(lk.viewInLobster)
-	//r.Path("/{builder}/builds/{buildnum:[0-9]+}/").HandlerFunc(viewBuild)
-	//r.Path("/{builder}/builds/{buildnum}/test/{test_phase}/{test_name}").HandlerFunc(app.MakeHandler(Name("view_test")))
 	r.Path("/status").Methods("GET").HandlerFunc(lk.checkAppHealth)
 
 	return r
